requirements.txt
mongo==0.2.0
paho-mqtt==1.3.1
pymongo==3.5.1
